deploy-notifications can find database address from a link
By (optionally) consuming a database link, deployment manifests
do not need to specify explicit database addresses in their
manifest, which allows the manifest to be simpler and more
automatable.

This currently only looks for address because that is the
only property consistent in database links. When database links
begin providing more-comprehensive properties, the consumption
can be updated.

This builds on #14